Re-organize util functions
